+++ CHANGELOG.md
@@ -6,9 +6,9 @@ You can also check [on GitHub](https://github.com/nextcloud/news/releases), the 
 
 # Unreleased
 ### Changed
+- Implement the cron check as a setup check (#2831)
 
 ### Fixed
-- Implement the cron check as a setup check (#2831)
 
 # Releases
 ## [28.0.0] - 2026-03-03